refactor token v2 models
