+++ rust/processor/src/db/common/models/mod.rs
@@ -5,4 +5,7 @@ pub mod event_models;
 pub mod fungible_asset_models;
 pub mod object_models;
 pub mod stake_models;
+pub mod token_models;
 pub mod token_v2_models;
+
+const DEFAULT_NONE: &str = "NULL";
