+++ rust/processor/src/db/common/models/token_models/mod.rs
@@ -1,11 +1,7 @@
-// Copyright © Aptos Foundation
-// SPDX-License-Identifier: Apache-2.0
-
 pub mod collection_datas;
-pub mod nft_points;
-pub mod token_activities;
 pub mod token_claims;
 pub mod token_datas;
 pub mod token_ownerships;
+pub mod token_royalty;
 pub mod token_utils;
 pub mod tokens;
