+++ rust/processor/src/db/common/models/token_v2_models/v2_token_datas.rs
@@ -6,23 +6,31 @@
 #![allow(clippy::unused_unit)]
 
 use crate::{
+    bq_analytics::generic_parquet_processor::{GetTimeStamp, HasVersion, NamedTable},
     db::{
         common::models::{
             object_models::v2_object_utils::ObjectAggregatedDataMapping,
+            token_models::token_utils::TokenWriteSet,
             token_v2_models::v2_token_utils::{TokenStandard, TokenV2, TokenV2Burned},
+            DEFAULT_NONE,
         },
-        postgres::models::{
-            resources::FromWriteResource, token_models::token_utils::TokenWriteSet,
-        },
+        postgres::models::resources::FromWriteResource,
     },
+    schema::current_token_datas_v2,
     utils::util::standardize_address,
 };
+use allocative_derive::Allocative;
+use anyhow::Context;
 use aptos_protos::transaction::v1::{DeleteResource, WriteResource, WriteTableItem};
 use bigdecimal::BigDecimal;
+use diesel::prelude::*;
+use field_count::FieldCount;
+use parquet_derive::ParquetRecordWriter;
 use serde::{Deserialize, Serialize};
+use tracing::error;
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct RawTokenDataV2 {
+pub struct TokenDataV2 {
     pub transaction_version: i64,
     pub write_set_change_index: i64,
     pub token_data_id: String,
@@ -43,12 +51,8 @@ pub struct RawTokenDataV2 {
     pub is_deleted_v2: Option<bool>,
 }
 
-pub trait TokenDataV2Convertible {
-    fn from_raw(raw_item: RawTokenDataV2) -> Self;
-}
-
 #[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct RawCurrentTokenDataV2 {
+pub struct CurrentTokenDataV2 {
     pub token_data_id: String,
     pub collection_id: String,
     pub token_name: String,
@@ -68,10 +72,10 @@ pub struct RawCurrentTokenDataV2 {
 }
 
 pub trait CurrentTokenDataV2Convertible {
-    fn from_raw(raw_item: RawCurrentTokenDataV2) -> Self;
+    fn from_raw(raw_item: CurrentTokenDataV2) -> Self;
 }
 
-impl RawTokenDataV2 {
+impl TokenDataV2 {
     // TODO: remove the useless_asref lint when new clippy nighly is released.
     #[allow(clippy::useless_asref)]
     pub fn get_v2_from_write_resource(
@@ -80,7 +84,7 @@ impl RawTokenDataV2 {
         write_set_change_index: i64,
         txn_timestamp: chrono::NaiveDateTime,
         object_metadatas: &ObjectAggregatedDataMapping,
-    ) -> anyhow::Result<Option<(Self, RawCurrentTokenDataV2)>> {
+    ) -> anyhow::Result<Option<(Self, CurrentTokenDataV2)>> {
         if let Some(inner) = &TokenV2::from_write_resource(write_resource)? {
             let token_data_id = standardize_address(&write_resource.address.to_string());
             let mut token_name = inner.get_name_trunc();
@@ -130,7 +134,7 @@ impl RawTokenDataV2 {
                     decimals: None,
                     is_deleted_v2: None,
                 },
-                RawCurrentTokenDataV2 {
+                CurrentTokenDataV2 {
                     token_data_id,
                     collection_id,
                     token_name,
@@ -159,11 +163,11 @@ impl RawTokenDataV2 {
         txn_version: i64,
         txn_timestamp: chrono::NaiveDateTime,
         tokens_burned: &TokenV2Burned,
-    ) -> anyhow::Result<Option<RawCurrentTokenDataV2>> {
+    ) -> anyhow::Result<Option<CurrentTokenDataV2>> {
         let token_data_id = standardize_address(&write_resource.address.to_string());
         // reminder that v1 events won't get to this codepath
         if let Some(burn_event_v2) = tokens_burned.get(&standardize_address(&token_data_id)) {
-            Ok(Some(RawCurrentTokenDataV2 {
+            Ok(Some(CurrentTokenDataV2 {
                 token_data_id,
                 collection_id: burn_event_v2.get_collection_address(),
                 token_name: "".to_string(),
@@ -191,11 +195,11 @@ impl RawTokenDataV2 {
         txn_version: i64,
         txn_timestamp: chrono::NaiveDateTime,
         tokens_burned: &TokenV2Burned,
-    ) -> anyhow::Result<Option<RawCurrentTokenDataV2>> {
+    ) -> anyhow::Result<Option<CurrentTokenDataV2>> {
         let token_data_id = standardize_address(&delete_resource.address.to_string());
         // reminder that v1 events won't get to this codepath
         if let Some(burn_event_v2) = tokens_burned.get(&standardize_address(&token_data_id)) {
-            Ok(Some(RawCurrentTokenDataV2 {
+            Ok(Some(CurrentTokenDataV2 {
                 token_data_id,
                 collection_id: burn_event_v2.get_collection_address(),
                 token_name: "".to_string(),
@@ -222,7 +226,7 @@ impl RawTokenDataV2 {
         txn_version: i64,
         write_set_change_index: i64,
         txn_timestamp: chrono::NaiveDateTime,
-    ) -> anyhow::Result<Option<(Self, RawCurrentTokenDataV2)>> {
+    ) -> anyhow::Result<Option<(Self, CurrentTokenDataV2)>> {
         let table_item_data = table_item.data.as_ref().unwrap();
 
         let maybe_token_data = match TokenWriteSet::from_table_item_type(
@@ -270,7 +274,7 @@ impl RawTokenDataV2 {
                         decimals: None,
                         is_deleted_v2: None,
                     },
-                    RawCurrentTokenDataV2 {
+                    CurrentTokenDataV2 {
                         token_data_id,
                         collection_id,
                         token_name,
@@ -298,5 +302,183 @@ impl RawTokenDataV2 {
             }
         }
         Ok(None)
+    }
+}
+
+/// This is a parquet version of TokenDataV2
+
+#[derive(Allocative, Clone, Debug, Default, Deserialize, ParquetRecordWriter, Serialize)]
+pub struct ParquetTokenDataV2 {
+    pub txn_version: i64,
+    pub write_set_change_index: i64,
+    pub token_data_id: String,
+    pub collection_id: String,
+    pub token_name: String,
+    pub largest_property_version_v1: Option<String>, // String format of BigDecimal
+    pub token_uri: String,
+    pub token_properties: String,
+    pub description: String,
+    pub token_standard: String,
+    pub is_fungible_v2: Option<bool>,
+    #[allocative(skip)]
+    pub block_timestamp: chrono::NaiveDateTime,
+    pub is_deleted_v2: Option<bool>,
+}
+
+impl NamedTable for ParquetTokenDataV2 {
+    const TABLE_NAME: &'static str = "token_datas_v2";
+}
+
+impl HasVersion for ParquetTokenDataV2 {
+    fn version(&self) -> i64 {
+        self.txn_version
+    }
+}
+
+impl GetTimeStamp for ParquetTokenDataV2 {
+    fn get_timestamp(&self) -> chrono::NaiveDateTime {
+        self.block_timestamp
+    }
+}
+
+impl From<TokenDataV2> for ParquetTokenDataV2 {
+    fn from(raw_item: TokenDataV2) -> Self {
+        Self {
+            txn_version: raw_item.transaction_version,
+            write_set_change_index: raw_item.write_set_change_index,
+            token_data_id: raw_item.token_data_id,
+            collection_id: raw_item.collection_id,
+            token_name: raw_item.token_name,
+            largest_property_version_v1: raw_item
+                .largest_property_version_v1
+                .map(|v| v.to_string()),
+            token_uri: raw_item.token_uri,
+            token_properties: canonical_json::to_string(&raw_item.token_properties.clone())
+                .context("Failed to serialize token properties")
+                .unwrap(),
+            description: raw_item.description,
+            token_standard: raw_item.token_standard,
+            is_fungible_v2: raw_item.is_fungible_v2,
+            block_timestamp: raw_item.transaction_timestamp,
+            is_deleted_v2: raw_item.is_deleted_v2,
+        }
+    }
+}
+
+#[derive(Allocative, Clone, Debug, Default, Deserialize, ParquetRecordWriter, Serialize)]
+pub struct ParquetCurrentTokenDataV2 {
+    pub token_data_id: String,
+    pub collection_id: String,
+    pub token_name: String,
+    pub maximum: Option<String>,                     // BigDecimal
+    pub supply: Option<String>,                      // BigDecimal
+    pub largest_property_version_v1: Option<String>, // String format of BigDecimal
+    pub token_uri: String,
+    pub token_properties: String, // serde_json::Value,
+    pub description: String,
+    pub token_standard: String,
+    pub is_fungible_v2: Option<bool>,
+    pub last_transaction_version: i64,
+    #[allocative(skip)]
+    pub last_transaction_timestamp: chrono::NaiveDateTime,
+    // Deprecated, but still here for backwards compatibility
+    pub decimals: Option<i64>,
+    pub is_deleted_v2: Option<bool>,
+}
+
+impl NamedTable for ParquetCurrentTokenDataV2 {
+    const TABLE_NAME: &'static str = "current_token_datas_v2";
+}
+
+impl HasVersion for ParquetCurrentTokenDataV2 {
+    fn version(&self) -> i64 {
+        self.last_transaction_version
+    }
+}
+
+impl GetTimeStamp for ParquetCurrentTokenDataV2 {
+    fn get_timestamp(&self) -> chrono::NaiveDateTime {
+        self.last_transaction_timestamp
+    }
+}
+
+impl From<CurrentTokenDataV2> for ParquetCurrentTokenDataV2 {
+    fn from(raw_item: CurrentTokenDataV2) -> Self {
+        Self {
+            token_data_id: raw_item.token_data_id,
+            collection_id: raw_item.collection_id,
+            token_name: raw_item.token_name,
+            maximum: raw_item.maximum.map(|v| v.to_string()),
+            supply: raw_item.supply.map(|v| v.to_string()),
+            largest_property_version_v1: raw_item
+                .largest_property_version_v1
+                .map(|v| v.to_string()),
+            token_uri: raw_item.token_uri,
+            token_properties: canonical_json::to_string(&raw_item.token_properties).unwrap_or_else(
+                |_| {
+                    error!(
+                        "Failed to serialize token_properties to JSON: {:?}",
+                        raw_item.token_properties
+                    );
+                    DEFAULT_NONE.to_string()
+                },
+            ),
+            description: raw_item.description,
+            token_standard: raw_item.token_standard,
+            is_fungible_v2: raw_item.is_fungible_v2,
+            last_transaction_version: raw_item.last_transaction_version,
+            last_transaction_timestamp: raw_item.last_transaction_timestamp,
+            decimals: raw_item.decimals,
+            is_deleted_v2: raw_item.is_deleted_v2,
+        }
+    }
+}
+
+/// This is a postgres version of TokenDataV2
+
+// PK of current_token_datas_v2, i.e. token_data_id
+pub type CurrentTokenDataV2PK = String;
+
+#[derive(Clone, Debug, Deserialize, FieldCount, Identifiable, Insertable, Serialize)]
+#[diesel(primary_key(token_data_id))]
+#[diesel(table_name = current_token_datas_v2)]
+pub struct PostgresCurrentTokenDataV2 {
+    pub token_data_id: String,
+    pub collection_id: String,
+    pub token_name: String,
+    pub maximum: Option<BigDecimal>,
+    pub supply: Option<BigDecimal>,
+    pub largest_property_version_v1: Option<BigDecimal>,
+    pub token_uri: String,
+    pub token_properties: serde_json::Value,
+    pub description: String,
+    pub token_standard: String,
+    pub is_fungible_v2: Option<bool>,
+    pub last_transaction_version: i64,
+    pub last_transaction_timestamp: chrono::NaiveDateTime,
+    // Deprecated, but still here for backwards compatibility
+    pub decimals: Option<i64>,
+    pub is_deleted_v2: Option<bool>,
+}
+
+impl From<CurrentTokenDataV2> for PostgresCurrentTokenDataV2 {
+    fn from(raw_item: CurrentTokenDataV2) -> Self {
+        Self {
+            token_data_id: raw_item.token_data_id,
+            collection_id: raw_item.collection_id,
+            token_name: raw_item.token_name,
+            maximum: raw_item.maximum,
+            supply: raw_item.supply,
+            largest_property_version_v1: raw_item.largest_property_version_v1,
+            token_uri: raw_item.token_uri,
+            token_properties: raw_item.token_properties,
+            description: raw_item.description,
+            token_standard: raw_item.token_standard,
+            is_fungible_v2: raw_item.is_fungible_v2,
+            last_transaction_version: raw_item.last_transaction_version,
+            last_transaction_timestamp: raw_item.last_transaction_timestamp,
+            decimals: raw_item.decimals,
+            is_deleted_v2: raw_item.is_deleted_v2,
+        }
     }
 }
