+++ rust/processor/src/db/parquet/models/mod.rs
@@ -5,7 +5,6 @@ pub mod event_models;
 pub mod fungible_asset_models;
 pub mod object_models;
 pub mod stake_models;
-pub mod token_v2_models;
 pub mod transaction_metadata_model;
 pub mod user_transaction_models;
 const DEFAULT_NONE: &str = "NULL";
