--- rust/processor/src/db/parquet/models/token_v2_models/mod.rs
@@ -1,6 +1,0 @@
-pub mod token_claims;
-pub mod v1_token_royalty;
-pub mod v2_token_activities;
-pub mod v2_token_datas;
-pub mod v2_token_metadata;
-pub mod v2_token_ownerships;
