--- rust/processor/src/db/postgres/models/token_models/token_claims.rs
@@ -1,75 +1,0 @@
-// Copyright © Aptos Foundation
-// SPDX-License-Identifier: Apache-2.0
-
-// This is required because a diesel macro makes clippy sad
-#![allow(clippy::extra_unused_lifetimes)]
-#![allow(clippy::unused_unit)]
-
-use crate::{
-    db::common::models::token_v2_models::raw_token_claims::{
-        CurrentTokenPendingClaimConvertible, RawCurrentTokenPendingClaim,
-    },
-    schema::current_token_pending_claims,
-};
-use bigdecimal::BigDecimal;
-use field_count::FieldCount;
-use serde::{Deserialize, Serialize};
-
-#[derive(
-    Clone, Debug, Deserialize, Eq, FieldCount, Identifiable, Insertable, PartialEq, Serialize,
-)]
-#[diesel(primary_key(token_data_id_hash, property_version, from_address, to_address))]
-#[diesel(table_name = current_token_pending_claims)]
-pub struct CurrentTokenPendingClaim {
-    pub token_data_id_hash: String,
-    pub property_version: BigDecimal,
-    pub from_address: String,
-    pub to_address: String,
-    pub collection_data_id_hash: String,
-    pub creator_address: String,
-    pub collection_name: String,
-    pub name: String,
-    pub amount: BigDecimal,
-    pub table_handle: String,
-    pub last_transaction_version: i64,
-    pub last_transaction_timestamp: chrono::NaiveDateTime,
-    pub token_data_id: String,
-    pub collection_id: String,
-}
-
-impl Ord for CurrentTokenPendingClaim {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.token_data_id_hash
-            .cmp(&other.token_data_id_hash)
-            .then(self.property_version.cmp(&other.property_version))
-            .then(self.from_address.cmp(&other.from_address))
-            .then(self.to_address.cmp(&other.to_address))
-    }
-}
-
-impl PartialOrd for CurrentTokenPendingClaim {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl CurrentTokenPendingClaimConvertible for CurrentTokenPendingClaim {
-    fn from_raw(raw_item: RawCurrentTokenPendingClaim) -> Self {
-        Self {
-            token_data_id_hash: raw_item.token_data_id_hash,
-            property_version: raw_item.property_version,
-            from_address: raw_item.from_address,
-            to_address: raw_item.to_address,
-            collection_data_id_hash: raw_item.collection_data_id_hash,
-            creator_address: raw_item.creator_address,
-            collection_name: raw_item.collection_name,
-            name: raw_item.name,
-            amount: raw_item.amount,
-            table_handle: raw_item.table_handle,
-            last_transaction_version: raw_item.last_transaction_version,
-            last_transaction_timestamp: raw_item.last_transaction_timestamp,
-            token_data_id: raw_item.token_data_id,
-            collection_id: raw_item.collection_id,
-        }
-    }
-}
