--- rust/processor/src/db/postgres/models/token_v2_models/mod.rs
@@ -1,9 +1,0 @@
-// Copyright © Aptos Foundation
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod v1_token_royalty;
-pub mod v2_collections;
-pub mod v2_token_activities;
-pub mod v2_token_datas;
-pub mod v2_token_metadata;
-pub mod v2_token_ownerships;
